+++ RELEASE_NOTES.md
@@ -1,5 +1,6 @@
 - Add static class `Atomic` containing atomic integer operation intrinsics
 - Add support for fixed-size arrays in storage buffers
+- Fix computation of storage buffer access (field and property access, access in utility functions)
 
 ### 5.7.17
 - Improve handling of functions with unit arguments